Relocate files into their proper directories
