+++ src/containers/BooksList.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import Book from './Book';
+import Book from '../components/Book';
 
 const BooksList = ({ books }) => (
   <table>
